still some mistakes in config.cpp that need to be fixed
