+++ config_file/MacOS.conf
@@ -1,6 +1,6 @@
 server {
 
-listen [::]:8080 default_server;
+listen [::]:8080;
 
 server_name www.example.com;
 
@@ -25,23 +25,23 @@ location / {
 }
 
 location /redirect {
-    return 302 https:://eu.siteground.com/kb/domain-redirects/;
+    redirect 302 https:://eu.siteground.com/kb/domain-redirects/;
 }
 
-location /cgi-bin/*.py {
+location /cgi-bin {
     allowed_methods GET,POST;
-    index index.py
+    index index.py;
     cgi.py /opt/homebrew/bin/python3;
 }
 
 
-location /cgi-bin/*.php {
+location /cgi-bin {
     allowed_methods GET, POST, DELETE;
-    index index.php
+    index index.php;
     cgi .py /opt/homebrew/bin/php;
 }
-
 }
+
 
 server {
     listen 80;
@@ -50,3 +50,4 @@ server {
     index index.html;
     return 302 https://en.wikipedia.org/wiki/42_(number);
 }
+;(if nothing is defined we do the default setting, make a rule if not defined then throw error reject this server)
