+++ src/config_parser_tester/ConfigTester.cpp
@@ -1,29 +1,36 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   tester.cpp                                         :+:    :+:            */
+/*   ConfigTester.cpp                                   :+:    :+:            */
 /*                                                     +:+                    */
 /*   By: lizhang <lizhang@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/06/29 12:39:16 by lizhang       #+#    #+#                 */
-/*   Updated: 2023/06/29 16:16:34 by lizhang       ########   odam.nl         */
+/*   Updated: 2023/07/25 14:52:42 by lizhang       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 
-#include "../../include/config.hpp"
-#include <iostream>
-
+#include "../../include/Config.hpp"
+#include "ServerTester.hpp"
 
 //config file tester
-int main(void)
+int main(int ac, char **av)
 {
-	Config_mac Config("../../config_file/MacOS.conf");
+	if (ac != 2)
+	{
+		std::cout<<"please give one argument: file path"<<std::endl;
+		return (0);
+	}
+	std::string name(av[1]);
+	Config Conf(name);
 
-	std::cout<<"server name: "<<Config.get_server_name()<<std::endl;
-	std::cout<<"listen port: "<<Config.get_port()<<std::endl;
-	std::cout<<"listen port name: "<<Config.get_port_name()<<std::endl;
-	std::cout<<"root location: "<<Config.get_root()<<std::endl;
-	std::cout<<"index: "<<Config.get_index()<<std::endl;
+	std::vector<Server>	listServers;
+
+	listServers = Conf.getServers();
+	for(unsigned int i = 0; i < listServers.size(); i++)
+	{
+		testServer(listServers[i]);
+	}
 	return(0);
 }
